SSL added
SSL has been added and appears to be working correctly.
Wireshark was used to capture packets flowing to the server (both HTTP
and HTTPS) I was able to locate the information inside of the HTTP
packets, but not the HTTPS packets. While this is by no means exhaustive
testing, it does lead me to believe that user information is being
encrypted correctly. Testing was done through the app, and all features
still appear to be functional.

Currently we are using a self signed certificate for testing.